+++ .gitignore
@@ -67,3 +67,4 @@ fastlane/readme.md
 app/.cxx
 
 beaengine
+beaengine/
