+++ androidTest/src/androidTest/java/com/chickenhook/vmcracker/RegressionTest.kt
@@ -1,4 +1,4 @@
-package com.self.vmcracker
+package com.chickenhook.vmcracker
 
 import android.os.SystemClock
 import androidx.test.espresso.Espresso
@@ -9,6 +9,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
+import org.chickenhook.demoapp.MainActivity
 
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -22,7 +23,8 @@ import org.junit.Rule
 @RunWith(AndroidJUnit4::class)
 class RegressionTest {
     @get:Rule
-    var mActivityRule: ActivityTestRule<MainActivity> = ActivityTestRule(MainActivity::class.java)
+    var mActivityRule: ActivityTestRule<MainActivity> = ActivityTestRule(
+        MainActivity::class.java)
 
     @Test
     fun startMainActivity() {
