+++ androidTest/src/main/res/layout/activity_web_view.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WebViewActivity">
+    tools:context="org.chickenhook.demoapp.WebViewActivity">
 
     <WebView
-        android:id="@+id/webview"
+        android:id="@+id/chicken_web_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </LinearLayout>
