+++ androidTest/src/test/java/com/chickenhook/vmcracker/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.self.vmcracker
+package com.chickenhook.vmcracker
 
 import org.junit.Test
